+++ packages/hightide-native/src/global-contexts/hightide-config/HightideConfigUtils.ts
@@ -6,8 +6,18 @@ import type { HightideTheme } from '../../theme/types/theme'
 
 export const HightideConfigUtils = {
   defaultSupportedLocales: {
-    'de-DE': { localName: 'Deutsch' },
-    'en-US': { localName: 'English (US)' },
+    'de-DE': {
+      localName: 'Deutsch',
+      defaultIs24HourFormat: true,
+      defaultStartingWeekday: 'monday',
+      defaultCalendarType: 'Gregorian',
+    },
+    'en-US': {
+      localName: 'English (US)',
+      defaultIs24HourFormat: false,
+      defaultStartingWeekday: 'sunday',
+      defaultCalendarType: 'Gregorian',
+    },
   } as const satisfies SupportedLocalesConfig,
   defaultSupportedThemes: {
     light: {