+++ packages/hightide-native/src/global-contexts/localization/index.ts
@@ -7,4 +7,6 @@ export {
   type LocalizationContextValue,
   type LocaleInformation,
   type SupportedLocalesConfig,
+  type Weekday,
+  type CalendarType,
 } from '@helpwave/hightide-utils/context/localization'