+++ packages/hightide-web/src/global-contexts/localization/forward-exports.ts
@@ -6,4 +6,6 @@ export {
   type LocalizationContextValue,
   type LocaleInformation,
   type SupportedLocalesConfig,
+  type Weekday,
+  type CalendarType,
 } from '@helpwave/hightide-utils/context/localization'